New runner format.
* Inheritence to reduce redundant code.
